1. abstract Create/Delete Virtualcluster functions, and encapsulate them into a new interface MasterProvisioner 2. adapt existing minikube implementation to MasterProvisioner interface
